share implementation of in memory store between examples and integration tests
